+++ packages/google_mobile_ads/example/ios/Runner/AppDelegate.m
@@ -15,6 +15,7 @@
 #include "AppDelegate.h"
 #import "FLTGoogleMobileAdsPlugin.h"
 #include "GeneratedPluginRegistrant.h"
+@import webview_flutter_wkwebview;
 
 @interface NativeAdFactoryExample : NSObject <FLTNativeAdFactory>
 @end